add .json to the list of resolved extensions by webpack (#341)
### 🔦 Summary

Currently, you have to explicitly add a `.json` extension when you import a json file. Webpack 4 and Node resolve files with a `.json` extension even if it wasn't explicitly added.

This PR adds `.json` to the list of extensions that are resolved automatically.

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan

Try implicitly import a json file without adding a `.json` extension.