wait for app-server port before finishing app-server task (#538)
<!---
Thanks for submitting a pull request 😄 !
-->

<!--- If you're changing public APIs make sure to document them (README, FAQ) -->

<!--- Be as descriptive as possible when explaining what was changed. Link to an issue if one exists -->
### 🔦 Summary
Fixes #523. As discussed, this may be a breaking change for projects that ignore the PORT env variable, because we will wait forever on a port that is not going to be used. (Did any one say framework? 😜 ).

<!--- Describe how the proposed changes are tested -->
### 🗺️ Test Plan
Added a new test to the start command that makes sure that the message is printed only when the port is available.